+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Handle casts involving classes and interfaces in `PlatformDependentCastRule`.
 - Update rule description for `SwallowedExceptionsRule`.
 
 ## [0.30.1] - 2021-04-06